added data connectivity test

+++ src/main/java/com/mkcollege/dpdd_year2_student_management_system/data/dto/StudentDetailsDto.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 
 public class StudentDetailsDto {
 
-    private int studentNumber;
+    private Long studentNumber;
 
     private String firstName;
